Updated the admin view
Admin view is now styled and has a logout button that works. The admin controller has a showView() function to bring the user to login screen. The saved events view has been changed to make the navigation bar highlight the selected location. In freefoodmainview view i changed the navigation bar "Saved Events" to "My Events" to stay consistent with the other views.